Losing bits on request.cmd_flags
In a few places a temporary value smaller than a cmd_flags
is used to test for bits and or build up a new cmd_flags.

Change to use explict u64 values where appropriate.

This patch is place holder for: Mike Christie's separate operations ... series
https://lkml.kernel.org/r/1465155145-10812-1-git-send-email-mchristi@redhat.com

Once Mike's patches are stablized this patch can be dropped.

+++ block/blk-merge.c
@@ -604,7 +604,7 @@ static int ll_merge_requests_fn(struct request_queue *q, struct request *req,
  */
 void blk_rq_set_mixed_merge(struct request *rq)
 {
-	unsigned int ff = rq->cmd_flags & REQ_FAILFAST_MASK;
+	u64 ff = rq->cmd_flags & REQ_FAILFAST_MASK;
 	struct bio *bio;
 
 	if (rq->cmd_flags & REQ_MIXED_MERGE)
